# A few more optimizations