- Switch to a growth factor of 2x instead of 1.5x. This results in
increased performance at the cost of mildly increased memory usage.
- Add a fastpath for encoding ascii strings (which are common).