@jcrist 2021-07-02
Hmmm, once again the change to a 2x factor initially benchmarked faster, but later on benchmarked as no change (except for a slight memory increase). Reverted again. I suspect this has something to do with load on my testing machine.

The unicode change is definitely worth it though.