@jcrist 2021-07-02
Running the existing benchmarks:

**Master**

```
$ python bench.py
Benchmark - 1 object (validate=False)
- orjson...
  dumps: 0.24 us
  loads: 0.56 us
- msgpack...
  dumps: 0.31 us
  loads: 0.64 us
- msgspec...
  dumps: 0.17 us
  loads: 0.33 us
- pyrobuf...
  dumps: 0.69 us
  loads: 0.91 us
Benchmark - 1k objects (validate=False)
- orjson...
  dumps: 361.02 us
  loads: 1235.96 us
- msgpack...
  dumps: 584.22 us
  loads: 1482.81 us
- msgspec...
  dumps: 265.37 us
  loads: 832.29 us
- pyrobuf...
  dumps: 1075.35 us
  loads: 3965.58 us
```

**This PR**

```
Benchmark - 1 object (validate=False)
- orjson...
  dumps: 0.24 us
  loads: 0.66 us
- msgpack...
  dumps: 0.32 us
  loads: 0.65 us
- msgspec...
  dumps: 0.13 us
  loads: 0.31 us
- pyrobuf...
  dumps: 0.66 us
  loads: 0.89 us
Benchmark - 1k objects (validate=False)
- orjson...
  dumps: 343.97 us
  loads: 1309.50 us
- msgpack...
  dumps: 584.68 us
  loads: 1746.03 us
- msgspec...
  dumps: 220.07 us
  loads: 792.90 us
- pyrobuf...
  dumps: 1014.19 us
  loads: 3713.43 us
```